Removed FadingEdgeScrollView if the platform is web. (#37)
* Removed FadingEdgeScrollView if the platform is web.
Didn't investigate the problem further. Also updated the plugin thought maybe this would solve the problem but it didn't :/ (revert this change if there are any problems with over platforms now with this version of the plugin)
Somehow the FadingEdgeScrollView is broken for web and throws an UnimplementedError. (Someone should open an issue in the FadingEdgeScrollView repo)

* Make code shorter

Co-authored-by: Uli Prantz <prantzuli@gmail.com>
Co-authored-by: Marcel Garus <marcel.garus@gmail.com>